Decouple nodes
This change also rename KeepNodeActiveRequest to
NodeHeartbeat and extracts ChaindexingNodeTasksRunner
