+++ chaindexing/src/nodes/node_task.rs
@@ -1,5 +1,4 @@
 use std::sync::Arc;
-
 use tokio::sync::Mutex;
 
 #[derive(Clone)]
