+++ chaindexing/src/nodes/node_tasks_runner.rs
@@ -1,0 +1,6 @@
+use super::node_task::NodeTask;
+
+#[async_trait::async_trait]
+pub trait NodeTasksRunner {
+    async fn run(&self) -> Vec<NodeTask>;
+}
